+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include README.md
+include requirements.txt
