+++ README.md
@@ -1,4 +1,33 @@
-# blobtools
-Application for the visualisation of (draft) genome assemblies using TAGC (Taxon-annotated Gc-Coverage) plots
+blobtools
+===============================
 
-For the documentation, please refer to https://blobtools.readme.io/
+version number: 1.0
+author: Dominik R Laetsch
+
+Overview
+--------
+
+A modular command-line solution for visualisation, quality control and taxonomic partitioning of genome datasets
+
+Installation / Usage
+--------------------
+
+To install use pip:
+
+    $ pip install blobtools
+
+
+Or clone the repo:
+
+    $ git clone https://github.com/DRL/blobtools.git
+    $ python setup.py install
+    
+Contributing
+------------
+
+TBD
+
+Example
+-------
+
+TBD
