+++ blobtools
@@ -1,1 +1,3 @@
-blobtools.py
+#!/usr/bin/env bash
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+$DIR/lib/blobtools.py "$@"
