+++ requirements.txt
@@ -1,0 +1,3 @@
+docopt==0.6.2
+matplotlib==2.0.2
+ujson==1.35
