+++ setup.cfg
@@ -1,0 +1,5 @@
+[bdist_wheel]
+universal=1
+
+[metadata]
+description-file=README.md
