--- test_files/blast.out
@@ -1,16 +1,0 @@
-contig_1    979556  200
-contig_2    979556  500
-contig_2    979556  1000
-contig_2    979556  500
-contig_2    979556  300
-contig_3    979556  10000
-contig_4    979556  1000
-contig_5    6252    2000
-contig_6    232323  2000
-contig_6    6252    2000
-contig_6    979556  2000
-contig_6    232323  2000
-contig_7    6252    2000
-contig_8    6252    2000
-contig_8    979556  2000
-contig_9    6252    200
