--- test_files/hits_file.assembly.fna
@@ -1,23 +1,0 @@
->ECOLI.contig_16
-AGCTAGCTAGCTAGCTAGCTAGTCGATCAT
->ECOLI.contig_17
-AGCTAGCTAGCTAGCTAGCTAGTCGATCAT
->ECOLI.contig_18
-AGCTAGCTAGCTAGCTAGCTAGTCGATCAT
->ECOLI.contig_19
-AGCTAGCTAGCTAGCTAGCTAGTCGATCAT
->ECOLI.contig_20
-AGCTAGCTAGCTAGCTAGCTAGTCGATCAT
->ECOLI.contig_21
-AGCTAGCTAGCTAGCTAGCTAGTCGATCAT
->ECOLI.contig_22
-AGCTAGCTAGCTAGCTAGCTAGTCGATCAT
->ECOLI.contig_23
-AGCTAGCTAGCTAGCTAGCTAGTCGATCAT
->ECOLI.contig_24
-AGCTAGCTAGCTAGCTAGCTAGTCGATCAT
->ECOLI.contig_25
-AGCTAGCTAGCTAGCTAGCTAGTCGATCAT
->ECOLI.contig_26
-AGCTAGCTAGCTAGCTAGCTAGTCGATCAT
-
