--- test_files/hits_file.blobDB.table.txt
@@ -1,22 +1,0 @@
-## blobtools v0.9.19
-## assembly	: /Users/dom/git/blobtools/test_files/hits_file.assembly.fna
-## coverage	: cov0 - /Users/dom/git/blobtools/test_files/hits_file.cov
-## taxonomy	: tax0 - /Users/dom/git/blobtools/test_files/hits_file.txt
-## nodesDB	: /Users/dom/git/blobtools/data/nodesDB.txt
-## taxrule	: bestsum
-## min_score	: 0.0
-## min_diff	: 0.0
-## tax_collision_random	: False
-##
-# name	length	GC	N	cov0	superkingdom.t.6	superkingdom.s.7	superkingdom.c.8	superkingdom.hits.9	phylum.t.10	phylum.s.11	phylum.c.12	phylum.hits.13	order.t.14	order.s.15	order.c.16	order.hits.17	family.t.18	family.s.19	family.c.20	family.hits.21	genus.t.22	genus.s.23	genus.c.24	genus.hits.25	species.t.26	species.s.27	species.c.28	species.hits.29
-ECOLI.contig_16	30	0.4667	0	100.1	Bacteria	3.29	0	tax0=Bacteria:3.29;	Proteobacteria	3.29	0	tax0=Proteobacteria:3.29;	Enterobacteriales	3.29	0	tax0=Enterobacteriales:3.29;	Enterobacteriaceae	3.29	0	tax0=Enterobacteriaceae:3.29;	Escherichia	3.29	0	tax0=Escherichia:3.29;	Escherichia coli	3.29	0	tax0=Escherichia coli:3.29;
-ECOLI.contig_17	30	0.4667	0	11.1	Bacteria	3.865	0	tax0=Bacteria:3.865;	Proteobacteria	3.865	0	tax0=Proteobacteria:3.865;	Enterobacteriales	3.865	0	tax0=Enterobacteriales:3.865;	Enterobacteriaceae	3.865	0	tax0=Enterobacteriaceae:3.865;	Escherichia	3.865	0	tax0=Escherichia:3.865;	Escherichia coli	3.865	0	tax0=Escherichia coli:3.865;
-ECOLI.contig_18	30	0.4667	0	11.1	Bacteria	61442.0	0	tax0=Bacteria:61442.0;	Proteobacteria	61442.0	0	tax0=Proteobacteria:61442.0;	Enterobacteriales	61442.0	0	tax0=Enterobacteriales:61442.0;	Enterobacteriaceae	61442.0	0	tax0=Enterobacteriaceae:61442.0;	Escherichia	61442.0	0	tax0=Escherichia:61442.0;	Escherichia coli	61442.0	0	tax0=Escherichia coli:61442.0;
-ECOLI.contig_19	30	0.4667	0	11.1	Bacteria	7424.0	0	tax0=Bacteria:7424.0;	Proteobacteria	7424.0	0	tax0=Proteobacteria:7424.0;	Enterobacteriales	7424.0	0	tax0=Enterobacteriales:7424.0;	Enterobacteriaceae	7424.0	0	tax0=Enterobacteriaceae:7424.0;	Escherichia	7424.0	0	tax0=Escherichia:7424.0;	Escherichia coli	7424.0	0	tax0=Escherichia coli:7424.0;
-ECOLI.contig_20	30	0.4667	0	11.1	Bacteria	1.045	0	tax0=Bacteria:1.045;	Proteobacteria	1.045	0	tax0=Proteobacteria:1.045;	Enterobacteriales	1.045	0	tax0=Enterobacteriales:1.045;	Enterobacteriaceae	1.045	0	tax0=Enterobacteriaceae:1.045;	Escherichia	1.045	0	tax0=Escherichia:1.045;	Escherichia coli	1.045	0	tax0=Escherichia coli:1.045;
-ECOLI.contig_21	30	0.4667	0	11.1	Bacteria	2.446	0	tax0=Bacteria:2.446;	Proteobacteria	2.446	0	tax0=Proteobacteria:2.446;	Enterobacteriales	2.446	0	tax0=Enterobacteriales:2.446;	Enterobacteriaceae	2.446	0	tax0=Enterobacteriaceae:2.446;	Escherichia	2.446	0	tax0=Escherichia:2.446;	Escherichia coli	2.446	0	tax0=Escherichia coli:2.446;
-ECOLI.contig_22	30	0.4667	0	11.1	Bacteria	58189.0	0	tax0=Bacteria:58189.0;	Proteobacteria	58189.0	0	tax0=Proteobacteria:58189.0;	Enterobacteriales	58189.0	0	tax0=Enterobacteriales:58189.0;	Enterobacteriaceae	58189.0	0	tax0=Enterobacteriaceae:58189.0;	Escherichia	58189.0	0	tax0=Escherichia:58189.0;	Escherichia coli	58189.0	0	tax0=Escherichia coli:58189.0;
-ECOLI.contig_23	30	0.4667	0	11.1	Bacteria	3.762	0	tax0=Bacteria:3.762;	Proteobacteria	3.762	0	tax0=Proteobacteria:3.762;	Enterobacteriales	3.762	0	tax0=Enterobacteriales:3.762;	Enterobacteriaceae	3.762	0	tax0=Enterobacteriaceae:3.762;	Escherichia	3.762	0	tax0=Escherichia:3.762;	Escherichia coli	3.762	0	tax0=Escherichia coli:3.762;
-ECOLI.contig_24	30	0.4667	0	11.1	Bacteria	4.367	0	tax0=Bacteria:4.367;	Proteobacteria	4.367	0	tax0=Proteobacteria:4.367;	Enterobacteriales	4.367	0	tax0=Enterobacteriales:4.367;	Enterobacteriaceae	4.367	0	tax0=Enterobacteriaceae:4.367;	Escherichia	4.367	0	tax0=Escherichia:4.367;	Escherichia coli	4.367	0	tax0=Escherichia coli:4.367;
-ECOLI.contig_25	30	0.4667	0	11.1	Bacteria	66258.2	0	tax0=Bacteria:66258.2;	Proteobacteria	66258.2	0	tax0=Proteobacteria:66258.2;	Enterobacteriales	66258.2	0	tax0=Enterobacteriales:66258.2;	Enterobacteriaceae	66258.2	0	tax0=Enterobacteriaceae:66258.2;	Escherichia	66258.2	0	tax0=Escherichia:66258.2;	Escherichia coli	66258.2	0	tax0=Escherichia coli:66258.2;
-ECOLI.contig_26	30	0.4667	0	11.1	Bacteria	1.449	0	tax0=Bacteria:1.449;	Proteobacteria	1.449	0	tax0=Proteobacteria:1.449;	Enterobacteriales	1.449	0	tax0=Enterobacteriales:1.449;	Enterobacteriaceae	1.449	0	tax0=Enterobacteriaceae:1.449;	Escherichia	1.449	0	tax0=Escherichia:1.449;	Escherichia coli	1.449	0	tax0=Escherichia coli:1.449;
