--- test_files/hits_file.txt
@@ -1,12 +1,0 @@
-ECOLI.contig_16 511145  3.290e+05       ECOLI.contig_16 CP009685.1      99.999  178162  1       0       1       178162  334588  512749  0.0
-ECOLI.contig_17 511145  3.865e+05       ECOLI.contig_17 CP014225.1      100.000 209298  0       0       1       209298  502194  711491  0.0
-ECOLI.contig_18 562     61442   ECOLI.contig_18 CP019629.1      100.000 33272   0       0       1       33272   316465  349736  0.0
-ECOLI.contig_19 562     742a    ECOLI.contig_19 CP013952.1      100.000 4020    0       0       3852    7871    1385454 1389473 0.0
-ECOLI.contig_20 562     1.045e+05       ECOLI.contig_20 CP013253.1      99.995  56618   3       0       1       56618   56833   216     0.0
-ECOLI.contig_21 511145  2.446e+05       ECOLI.contig_21 CP009685.1      100.000 132468  0       0       1       132468  3170043 3302510 0.0
-ECOLI.contig_22 562     58189   ECOLI.contig_22 CP013952.1      100.000 31510   0       0       1       31510   3784471 3815980 0.0
-ECOLI.contig_23 511145  3.762e+05       ECOLI.contig_23 CP009685.1      99.991  203791  19      0       1       203791  3675184 3471394 0.0
-ECOLI.contig_24 511145  4.367e+05       ECOLI.contig_24 CP009685.1      100.000 236482  0       0       1       236482  703778  940259  0.0
-ECOLI.contig_25 562     65258   ECOLI.contig_25 CP019629.1      99.929  35423   2       3       1       35402   1651531 1686951 0.0
-ECOLI.contig_25 562	1000.2	ECOLI.contig_25 CP019629.1      99.929  35423   2       3       1       35402   1651531 1686951 0.0
-ECOLI.contig_26 562     1.449e+05       ECOLI.contig_26 CP013952.1      99.975  78531   19      1       1       78531   763689  685160  0.0
