--- test_files/test.mapping_1.bam.cov
@@ -1,16 +1,0 @@
-## blobtools v0.9.19
-## Total Reads = 15313
-## Mapped Reads = 15313
-## Unmapped Reads = 0
-## Source(s) : /Users/dom/git/blobtools/test_files/mapping_1.bam
-# contig_id	read_cov	base_cov
-contig_1	369	90.406
-contig_2	844	168.409
-contig_3	188	43.761
-contig_4	2096	456.313
-contig_5	456	163.557
-contig_6	52	25.88
-contig_7	1005	52.312
-contig_8	1008	91.742
-contig_9	554	74.757
-contig_10	8741	310.634
